# docs/001-capacity-exchange-marketplace/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Devotion, Capacity Exchange API
  version: 0.1.0
  description: |
    Kontrak API untuk Devotion, marketplace subkontrak kapasitas konveksi.

    Aturan yang berlaku di seluruh kontrak:
    - Seluruh pesan galat berbahasa Indonesia dan dapat dikutip penguji dalam laporan.
    - Nilai uang berupa bilangan bulat rupiah (`integer`, format int64). Tidak ada pecahan.
    - Periode mingguan berupa tanggal Senin awal minggu (`date`), tanpa zona waktu.
    - Waktu kejadian berupa `date-time` berzona; klien mengonversi ke WIB saat menampilkan.
    - Paginasi memakai kursor tuple, bukan offset, agar urutan stabil antar halaman.
    - Autentikasi memakai cookie sesi `httpOnly`; tidak ada token di badan respons.

    Rujukan requirement ada pada setiap operasi lewat `x-fr`.
  license:
    name: MIT
servers:
  - url: https://{host}/api
    variables:
      host:
        default: localhost
        description: Satu origin; frontend disajikan proses yang sama.
...
security:
  - sessionCookie: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Daftar akun usaha baru
      security: []
      x-fr: [FR-001, FR-002]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Akun dibuat; kode verifikasi dikirim ke email dan nomor HP
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterResponse' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '409':
          description: Email atau nomor HP sudah terdaftar
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: 'https://devotion/errors/duplicate'
                title: Data sudah terdaftar
                status: 409
                code: EMAIL_ALREADY_REGISTERED
                detail: Email ini sudah dipakai akun lain. Masuk atau pulihkan kata sandi.
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verifikasi alamat email dengan kode enam digit
      security: []
      x-fr: [FR-002]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationCodeRequest' }
      responses:
        '204': { description: Email terverifikasi }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '410':
          description: Kode kedaluwarsa atau sudah dipakai
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /auth/verify-phone:
    post:
      tags: [auth]
      summary: Verifikasi nomor HP dengan kode enam digit dari WhatsApp
      security: []
      x-fr: [FR-002]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationCodeRequest' }
      responses:
        '204': { description: Nomor HP terverifikasi }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '410': { description: Kode kedaluwarsa atau sudah dipakai }
  /auth/resend-code:
    post:
      tags: [auth]
      summary: Kirim ulang kode verifikasi
      description: |
        Dibatasi 3 kali per jam per nomor dan 10 nomor berbeda per jam per alamat asal.
        Batas kedua menutup pemutaran nomor yang dapat memancing pemblokiran nomor pengirim.
      security: []
      x-fr: [FR-002]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target, channel]
              properties:
                target: { type: string, description: Email atau nomor HP }
                channel: { type: string, enum: [email, whatsapp] }
      responses:
        '202': { description: Permintaan diterima; pengiriman berjalan di latar }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
  /auth/login:
    post:
      tags: [auth]
      summary: Masuk dan buat sesi
      description: Dibatasi 5 percobaan per 15 menit per akun, dengan jeda yang membesar.
      security: []
      x-fr: [FR-003]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '200':
          description: Sesi dibuat; cookie sesi disetel
          headers:
            Set-Cookie:
              schema:
                type: string
                example: devotion_session=<opaque>; HttpOnly; Secure; SameSite=Lax; Path=/; Max-Age=604800
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyAccount' }
        '401':
          description: Email atau kata sandi salah
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: 'https://devotion/errors/invalid-credentials'
                title: Gagal masuk
                status: 401
                code: INVALID_CREDENTIALS
                detail: Email atau kata sandi salah.
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Keluar dan hapus sesi
      x-fr: [FR-003]
      responses:
        '204': { description: Sesi dihapus }
        '401': { $ref: '#/components/responses/NotAuthenticated' }
  /auth/recover/request:
    post:
      tags: [auth]
      summary: Minta kode pemulihan kata sandi
      description: Selalu menjawab 202 agar tidak membocorkan apakah sebuah email terdaftar.
      security: []
      x-fr: [FR-003]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202': { description: Bila akun ada, kode pemulihan dikirim }
  /auth/recover/confirm:
    post:
      tags: [auth]
      summary: Setel kata sandi baru dengan kode pemulihan
      security: []
      x-fr: [FR-003]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, new_password]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^[0-9]{6}$' }
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        '204': { description: Kata sandi diganti; seluruh sesi lain dihapus }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '410': { description: Kode kedaluwarsa atau sudah dipakai }
  /me:
    get:
      tags: [auth]
      summary: Akun yang sedang masuk
      x-fr: [FR-001, FR-005]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyAccount' }
        '401': { $ref: '#/components/responses/NotAuthenticated' }
  /me/roles:
    patch:
      tags: [auth]
      summary: Tambah peran usaha pada akun yang sama
      description: Peran dapat ditambahkan, tidak dapat dicabut sendiri bila masih ada pesanan aktif.
      x-fr: [FR-001]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subcontractor: { type: boolean }
                buyer: { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyAccount' }
        '409':
          description: Peran tidak dapat dicabut karena masih ada pesanan aktif
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ─────────────────────────── profile ───────────────────────────
  /profile/me:
    get:
      tags: [profile]
      summary: Profil usaha milik akun yang masuk
      x-fr: [FR-004]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyProfile' }
        '401': { $ref: '#/components/responses/NotAuthenticated' }
        '403':
          description: Akun admin tidak memiliki profil usaha
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: 'https://devotion/errors/forbidden'
                title: Tidak berwenang
                status: 403
                code: FORBIDDEN
                detail: Akun admin tidak memiliki profil usaha.
    put:
      tags: [profile]
      summary: Ubah profil usaha
      description: |
        Koordinat bersifat opsional dan tampil publik pada peta profil (FR-057, FR-064).
        Bila lintang diisi, bujur wajib diisi juga.
      x-fr: [FR-004, FR-057]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyProfile' }
        '401': { $ref: '#/components/responses/NotAuthenticated' }
        '403':
          description: Akun admin tidak memiliki profil usaha
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: 'https://devotion/errors/forbidden'
                title: Tidak berwenang
                status: 403
                code: FORBIDDEN
                detail: Akun admin tidak memiliki profil usaha.
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /profile/{profileId}:
    get:
      tags: [profile]
      summary: Profil publik sebuah usaha
      description: |
        Memuat atribut listing, ketersediaan kapasitas terkini, titik lokasi, dan ringkasan
        reputasi. Tingkat penyelesaian ditahan sebagai persentase sampai ada minimal 3
        pesanan yang disepakati (FR-073).
      security: []
      x-fr: [FR-016, FR-048, FR-064, FR-071, FR-073]
      parameters:
        - { name: profileId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProfile' }
        '404': { $ref: '#/components/responses/NotFound' }
  # ─────────────────────────── master data ───────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: devotion_session
      description: |
        Cookie sesi httpOnly, Secure, SameSite=Lax. Yang disimpan di server adalah hash
        token, bukan token mentah. Kirim dengan credentials 'include'.
  responses:
    ValidationFailed:
      description: Masukan tidak lolos validasi
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemValidation' }
          example:
            type: 'https://devotion/errors/validation'
            title: Masukan tidak sah
            status: 422
            code: VALIDATION_FAILED
            detail: Beberapa isian tidak sesuai ketentuan.
            errors:
              - { field: email, message: Format email tidak sah }
    NotAuthenticated:
      description: Belum masuk atau sesi tidak sah
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://devotion/errors/not-authenticated'
            title: Belum masuk
            status: 401
            code: NOT_AUTHENTICATED
            detail: Sesi tidak ditemukan atau telah berakhir. Silakan masuk kembali.
...
    NotFound:
      description: Sumber daya tidak ditemukan
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://devotion/errors/not-found'
            title: Tidak ditemukan
            status: 404
            code: NOT_FOUND
            detail: Data yang diminta tidak ditemukan.
    RateLimitExceeded:
      description: Terlalu banyak percobaan
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://devotion/errors/rate-limited'
            title: Terlalu banyak percobaan
            status: 429
            code: RATE_LIMIT_EXCEEDED
            detail: Terlalu banyak percobaan. Coba lagi beberapa saat lagi.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Detik hingga percobaan berikutnya diperbolehkan
  schemas:
    Problem:
      type: object
      description: Galat mengikuti RFC 9457 (application/problem+json).
      required: [type, title, status, code, detail]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Kode mesin yang stabil untuk penanganan di klien.
          enum:
            - VALIDATION_FAILED
            - NOT_AUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - RATE_LIMIT_EXCEEDED
            - EMAIL_ALREADY_REGISTERED
            - INVALID_CREDENTIALS
            - INVALID_VERIFICATION_CODE
            - VERIFICATION_CODE_EXPIRED
            - EMAIL_NOT_VERIFIED
            - PHONE_NOT_VERIFIED
            - IDENTITY_NOT_VERIFIED
            - IDENTITY_ALREADY_VERIFIED
            - VERIFICATION_PENDING
            - FILE_TOO_LARGE
            - UNSUPPORTED_FILE_TYPE
            - STORAGE_QUOTA_FULL
            - LISTING_NOT_FOUND
            - LISTING_ALREADY_EXISTS
            - CAPACITY_ALREADY_ALLOCATED
            - PERIOD_ALREADY_ALLOCATED
            - SELF_REQUEST
            - INSUFFICIENT_CAPACITY
            - REQUEST_EXPIRED
            - REQUEST_ALREADY_AGREED
            - CAPACITY_ALREADY_TAKEN
            - INVALID_STATUS_TRANSITION
            - CANCELLATION_AFTER_PRODUCTION
            - WORK_ORDER_NOT_COMPLETED
            - REVIEW_ALREADY_SUBMITTED
            - READINESS_AFTER_DEADLINE
        detail:
          type: string
          description: Penjelasan bahasa Indonesia yang dapat dikutip pengguna.
        instance: { type: string }
        meta: { type: object, additionalProperties: true }
    ProblemValidation:
      allOf:
        - { $ref: '#/components/schemas/Problem' }
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
...
    RegisterRequest:
      type: object
      description: |
        Pendaftaran sekaligus membuat profil usaha dalam satu transaksi, sehingga
        `city_code` wajib dan `GET /api/profile/me` tidak pernah 404. Minimal satu peran
        usaha dipilih saat mendaftar (FR-001).
      required: [email, phone, password, business_name, city_code, roles]
      properties:
        email: { type: string, format: email }
        phone: { type: string, description: Nomor telepon format E.164 Indonesia }
        password: { type: string, format: password, minLength: 8 }
        business_name: { type: string, minLength: 3, maxLength: 150 }
        city_code: { type: string, pattern: '^[0-9]{4}$', description: Kode kota/kabupaten tempat usaha }
        roles:
          type: object
          required: [subcontractor, buyer]
          description: Minimal satu bernilai true.
          properties:
            subcontractor: { type: boolean }
            buyer: { type: boolean }
    RegisterResponse:
      type: object
      required: [email_verification_required]
      properties:
        email_verification_required: { type: boolean }
        phone_verification_required: { type: boolean }
    VerificationCodeRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 4, maxLength: 8 }
    MyAccount:
      type: object
      required: [account_id, email, email_verified, phone_verified, roles]
      properties:
        account_id: { type: string, format: uuid }
        email: { type: string, format: email }
        phone: { type: string }
        email_verified: { type: boolean }
        phone_verified: { type: boolean }
        roles:
          type: object
          properties:
            subcontractor: { type: boolean }
            buyer: { type: boolean }
        profile_id: { type: string, format: uuid, nullable: true }
        is_admin: { type: boolean }

    # ── profile ──
    ProfileUpdateRequest:
      type: object
      description: |
        Provinsi tidak dikirim: ia turunan dari `city.province_code`. Batas alamat tidak
        disimpan, jadi tidak ada field `address`. Rentang koordinat menyalin constraint
        `coordinates_within_indonesia` agar koordinat di luar Indonesia menjadi 422, bukan 500.
      properties:
        business_name: { type: string, minLength: 3, maxLength: 150 }
        description: { type: string, maxLength: 2000 }
        city_code: { type: string, pattern: '^[0-9]{4}$' }
        latitude: { type: number, format: double, nullable: true, minimum: -11.5, maximum: 6.5 }
        longitude: { type: number, format: double, nullable: true, minimum: 94.5, maximum: 141.5 }
    MyProfile:
      type: object
      required: [profile_id, business_name, identity_verified]
      properties:
        profile_id: { type: string, format: uuid }
        business_name: { type: string }
        description: { type: string, nullable: true }
        city_code: { type: string, nullable: true }
        city_name: { type: string, nullable: true, readOnly: true }
        province_code: { type: string, nullable: true, readOnly: true }
        province_name: { type: string, nullable: true, readOnly: true }
        latitude: { type: number, format: double, nullable: true }
        longitude: { type: number, format: double, nullable: true }
        identity_verified: { type: boolean }
        verification_status: { allOf: [ { $ref: '#/components/schemas/VerificationStatus' } ], nullable: true }
        reputation: { $ref: '#/components/schemas/Reputation' }
    PublicProfile:
      type: object
      required: [profile_id, business_name, identity_verified]
      properties:
        profile_id: { type: string, format: uuid }
        business_name: { type: string }
        description: { type: string, nullable: true }
        city_code: { type: string, nullable: true }
        city_name: { type: string, nullable: true }
        province_code: { type: string, nullable: true }
        province_name: { type: string, nullable: true }
        latitude: { type: number, format: double, nullable: true }
        longitude: { type: number, format: double, nullable: true }
        identity_verified:
          type: boolean
          description: Status verifikasi identitas tidak memengaruhi skor pencarian (FR-034).
        listing: { allOf: [ { $ref: '#/components/schemas/Listing' } ], nullable: true }
        reputation: { $ref: '#/components/schemas/Reputation' }
...
    VerificationStatus:
      type: string
      enum: [pending, approved, rejected]
    # ── master data ──
...